+++ lib/actions/actions.ts
@@ -1,0 +1,7 @@
+'use server';
+
+import { revalidateTag } from 'next/cache';
+
+export default async function revalidateCacheTag(tag: string) {
+    revalidateTag(tag);
+}
